Tests/LibWeb/Text/input/HTML/shadow-realm-btoa.html

<!DOCTYPE html>
<script src="../include.js"></script>
<script>
    test(() => {
        const realm = new ShadowRealm();
        const result = realm.evaluate(`() => self.btoa('Hello, world!')`)();
        println(`Shadow realm evaluation returned: ${result}`);
    });
</script>
