Tests/LibWeb/Text/expected/HTML/shadow-realm-btoa.txt

Shadow realm evaluation returned: SGVsbG8sIHdvcmxkIQ==
